[bugfix] avoid infinite loop in construction of VirtualizedLocalFunction
When a VirtualizedLocalFunction was constructed from a
VirtualizedGridFunction, its own constructor was called recursively.

This patch breaks this infinite loop by asking the
grid function (gf) to create the local function itself by calling
gf.localFunction(). If a grid function is not able to do that, we fall
back to creating the local function directly (passing the grid function
to the constructor). The VirtualizedGridFunction must provide this new
method, of course.