add new GlobalMapper
The GlobalMapper is similar to the MCMG mapper in use, but constructs a
global numbering instead of a local one. It can be used to index distributed
arrays in such a manner that they are consistently numbered when communicated
to a single rank.

The goal of this implementation is using the Matplotlib for data generated by
distributed computations (e.g., through an IPython notebook).

Note: The implementation is currently untested.