simplify code generation
Always writing open* and close* to start and end a method is tedious.

This patch introduces objects modeling namespaces, classes, methods, etc.
These objects have subobjects themselves and can be emitted by a single call
to the source writer.

The basic idea is to model the C++ grammar (similar to the way UFL models
expressions). Only the SourceWriter shall be responsible for the final code
formatting. Once all statements can be written in this way, we can reason
about the C++ code before emitting it. Depending on the use, we might still
have information that is no longer available to the compiler to do, e.g.,
loop fusion.