replace classes "NamedCoefficient" and "NamedConstant" by functions "Constant" and "Coefficient"
There is no class "Constant" in UFL. As we only want to add a tag "name" to the
coefficient, we simple construct the coefficient and assign it a name.

Having functions "Coefficient" and "Constant" in our own UFL space helps us
customize the UFL to our needs. We can now freely add further tags not used by
the UFL.