[bugfix][integrands] don't register "setCoefficient" a second time
In case of renumbering, setCoefficient was is renamed to _setCoefficient
and a Python function is registered doing the renumbering. Now, if we
try to load the module a second time, this renaming has already
happened and we write the Python function into _setCoefficient. As
this Python function calls _setCoefficient, we end up in an infinite
recursion.

This patch fixes the issue by registering setCoefficient only, if the
Integrands do not have an attribute named "_renumbering".