reduce conversion magic to dune.ufl.GridFunction
Historically, there was a lot of code converting grid functions to
dune.ufl.GridFunction. As this is now done automatically by pybind11,
this patch removes these conversions.

Slight step back: The current version will not interpolate Python expressions
automatically as the old one did. This feature will be included in the future
without creating a "globalGridFunction".