extend use of asGridFunction
Usually, one type of grid function is special on the C++ side, e.g., the
DomainFunctionType for a dune-fem operator. With this patch, this type can
be passed to asGridFunction and virtualization is not applied in this case.
Moreover, asGridFunction will check whether a VirtualizedGridFunction is
accepted by the "apply" functor, so that user code no longer has to take
care of this.

In the special case of DiscreteFunction.interpolate, interpolating a
discrete function of the same type will silently call "assign" instead
of "interpolate" on the dune-fem side.