support named constants in integrals model
When passing a `NamedConstant` to an integrands model, you can now
access the constant by `model.<name>()` on the C++ side. The Python
bindings support this through a property.

For this to work, the name must be a valid C++ identifier. Moreover, the
name may not clash with any C++ identifier automatically generated by
the integrands model generator (e.g., `entity`).